Refactoring of datamodel_model.py
