+++ src/models/endpoint_model.py
@@ -20,9 +20,9 @@ import time
 import uuid
 from typing import Optional, List
 
-import dds_data
-from dds_data import DataEndpoint
-from dds_utils import getProperty, HOSTNAMES, PROCESS_NAMES, PIDS, ADDRESSES
+from dds_access import dds_data
+from dds_access.dds_data import DataEndpoint
+from dds_access.dds_utils import getProperty, HOSTNAMES, PROCESS_NAMES, PIDS, ADDRESSES
 from utils import EntityType
